+++ source/extensions/lightspeed.trex.layout.stagecraft/docs/CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.0] - 2024-03-13
+### Added
+- Global event for page changes and file picker management
 
 ## [1.1.1]
 ### Changed