+++ source/extensions/lightspeed.trex.project_wizard.window/docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 ﻿# Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.0] - 2024-04-12
+### Changed
+- Now has a single global instance initialized and destroyed in `extension.py`
+- Hide and show wizard functionality driven by file picker open and close events
+
 ## [1.1.1]
 ### Changed
 - Set Apache 2 license headers